+++ src/components/TeacherViews/LessonOverview/LessonOverview.view.tsx
@@ -49,7 +49,6 @@ import {
 } from "@/components/TeacherComponents/helpers/downloadAndShareHelpers/downloadsCopyright";
 import { ExpiringBanner } from "@/components/SharedComponents/ExpiringBanner";
 import LessonOverviewMediaClips from "@/components/TeacherComponents/LessonOverviewMediaClips";
-import lessonMediaClipsFixtures from "@/node-lib/curriculum-api-2023/fixtures/lessonMediaClips.fixture";
 import LessonOverviewDocPresentation from "@/components/TeacherComponents/LessonOverviewDocPresentation";
 import { TeacherNoteInline } from "@/components/TeacherComponents/TeacherNoteInline/TeacherNoteInline";
 
@@ -71,7 +70,7 @@ export const getDedupedPupilLessonOutcome = (
   }
   return plo;
 };
-export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
+export function LessonOverview({ lesson }: LessonOverviewProps) {
   const {
     lessonTitle,
     lessonSlug,
@@ -103,10 +102,13 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
     additionalMaterialUrl,
     actions,
     hasMediaClips,
+    lessonMediaClips,
     teacherNoteHtml,
     teacherNoteError,
     additionalFiles,
+    lessonOutline,
   } = lesson;
+
   const { track } = useAnalytics();
   const { analyticsUseCase } = useAnalyticsPageProps();
   const commonPathway = getPathway(lesson);
@@ -218,13 +220,17 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
 
   const showDownloadAll = downloadsFilteredByCopyright.length > 0;
   const showShare =
-    !isSpecialist && keyStageSlug !== "early-years-foundation-stage";
+    !isSpecialist &&
+    keyStageSlug !== "early-years-foundation-stage" &&
+    !actions?.disablePupilShare;
+
+  // TODO: use actions and exceptions for this
+  const isPELesson = subjectSlug === "physical-education";
 
   const isMFL =
     subjectSlug === "german" ||
     subjectSlug === "french" ||
-    subjectSlug === "spanish" ||
-    lessonSlug === "des-auteurs-francophones-perfect-tense-with-etre";
+    subjectSlug === "spanish";
 
   return (
     <MathJaxLessonProvider>
@@ -255,7 +261,7 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
         track={track}
         analyticsUseCase={analyticsUseCase}
         isNew={isNew}
-        isShareable={!expired}
+        isShareable={!expired && !actions?.disablePupilShare}
         onClickDownloadAll={() => {
           trackDownloadResourceButtonClicked({
             downloadResourceButtonName: "all",
@@ -380,8 +386,8 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
                     </LessonItemContainer>
                   )}
                 {pageLinks.find((p) => p.label === mediaClipLabel) &&
-                  hasMediaClips &&
-                  isBeta && (
+                  lessonMediaClips &&
+                  hasMediaClips && (
                     <LessonItemContainer
                       title={mediaClipLabel}
                       ref={lessonMediaClipsSectionRef}
@@ -390,14 +396,16 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
                       slugs={slugs}
                       pageLinks={pageLinks}
                       displayMediaClipButton={true}
+                      isCanonical={isCanonical}
                     >
                       <LessonOverviewMediaClips
                         lessonSlug={lessonSlug}
-                        learningCycleVideos={
-                          lessonMediaClipsFixtures().mediaClips
-                        }
+                        learningCycleVideos={lessonMediaClips}
+                        isCanonical={isCanonical}
                         unitSlug={unitSlug ?? null}
                         programmeSlug={programmeSlug ?? null}
+                        lessonOutline={lessonOutline}
+                        isPELesson={isPELesson}
                       />
                     </LessonItemContainer>
                   )}
@@ -424,7 +432,7 @@ export function LessonOverview({ lesson, isBeta }: LessonOverviewProps) {
                     isLegacyLicense={isLegacyLicense}
                     isMathJaxLesson={isMathJaxLesson}
                     hasVocabAndTranscripts={Boolean(additionalMaterialUrl)}
-                    displayVocab={isBeta && isMFL}
+                    displayVocab={isMFL}
                     updatedAt={updatedAt}
                     additionalFiles={additionalFiles}
                   />