+++ src/components/TeacherViews/SpecialistLesson/SpecialistLesson.stories.tsx
@@ -54,7 +54,9 @@ export const SpecialistLessonPage: Story = {
       updatedAt: "2022-02-22T12:00:00Z",
       lessonGuideUrl: null,
       hasMediaClips: false,
+      lessonMediaClips: null,
       additionalFiles: null,
+      lessonOutline: null,
     },
   },
 };