+++ src/pages/teachers/programmes/[programmeSlug]/units/[unitSlug]/lessons/[lessonSlug].tsx
@@ -165,7 +165,6 @@ export const getStaticProps: GetStaticProps<
       }
 
       const lessonPageData = await populateLessonWithTranscript(curriculumData);
-
       const results: GetStaticPropsResult<LessonOverviewPageProps> = {
         props: {
           curriculumData: lessonPageData,